+++ actfold/models/__init__.py
@@ -2,6 +2,15 @@
 
 from __future__ import annotations
 
+from actfold.models.architecture_utils import (
+    ArchitectureProfile,
+    ManualFoldedForward,
+    build_manual_folded_forward,
+    detect_architecture,
+    find_embedding_module,
+    find_layer_list,
+    find_lm_head,
+)
 from actfold.models.base import DiffusionLLM
 from actfold.models.diffusion_sampler import DiffusionSampler, SamplerConfig, SamplerOutput
 from actfold.models.dream_sampler import DreamSampler, DreamSamplerConfig
@@ -25,6 +34,7 @@ from actfold.models.sampling_utils import (
 )
 
 __all__ = [
+    "ArchitectureProfile",
     "CosineMaskingScheduler",
     "DiffusionLLM",
     "DiffusionSampler",
@@ -35,14 +45,20 @@ __all__ = [
     "LinearMaskingScheduler",
     "LLaDASampler",
     "LLaDASamplerConfig",
+    "ManualFoldedForward",
     "MaskingScheduler",
     "ModelRegistry",
     "SamplerConfig",
     "SamplerOutput",
     "add_gumbel_noise",
     "build_left_padded_canvas",
+    "build_manual_folded_forward",
     "build_right_padded_canvas",
     "compute_position_ids",
+    "detect_architecture",
+    "find_embedding_module",
+    "find_layer_list",
+    "find_lm_head",
     "get_num_transfer_tokens",
     "load_model",
     "make_masking_scheduler",